Add IT for mapping change events [WIP]

+++ mappingservice/lisp-proto/src/main/java/org/opendaylight/lispflowmapping/lisp/util/LispAddressUtil.java
@@ -704,7 +704,7 @@ public final class LispAddressUtil {
         return builder.build();
     }
 
-    public static Eid asSrcDstEid(String src, String dst, int smask, int dmask, int vni) {
+    public static Eid asSrcDstEid(String src, String dst, int smask, int dmask, long vni) {
         EidBuilder builder = new EidBuilder();
         builder.setAddressType(SourceDestKeyLcaf.class);
         builder.setVirtualNetworkId(new InstanceIdType(Long.valueOf(vni)));
